+++ app/src/main/java/com/example/dumb_app/core/model/Plan/PlanItemCreateDto.kt
@@ -4,5 +4,5 @@ data class PlanItemCreateDto(
     val type:    Int,
     val number:  Int,
     val tOrder:  Int,
-    val tWeight: Int
+    val tWeight: Float
 )
